feat(nodes): merge nginx templates into nodes
