--- nginx_app/apps.py
@@ -1,6 +1,0 @@
-from django.apps import AppConfig
-
-class NginxAppConfig(AppConfig):
-    default_auto_field = 'django.db.models.BigAutoField'
-    name = 'nginx_app'
-    verbose_name = 'NGINX Templates'
