+++ nodes/apps.py
@@ -4,3 +4,4 @@ from django.apps import AppConfig
 class NodesConfig(AppConfig):
     default_auto_field = "django.db.models.BigAutoField"
     name = "nodes"
+    verbose_name = "Node Infrastructure"
