+++ nodes/tests.py
@@ -1,12 +1,15 @@
 from pathlib import Path
 from unittest.mock import patch
 import socket
+import threading
+import http.server
+import socketserver
 
 from django.test import Client, TestCase
 from django.urls import reverse
 from django.contrib.auth import get_user_model
 
-from .models import Node, NodeScreenshot
+from .models import Node, NodeScreenshot, NginxConfig
 
 
 class NodeTests(TestCase):
@@ -62,4 +65,31 @@ class NodeAdminTests(TestCase):
         response = self.client.get(url)
         self.assertEqual(response.status_code, 302)
         self.assertEqual(Node.objects.count(), 1)
+
+
+class NginxConfigTests(TestCase):
+    def _run_server(self, port):
+        handler = http.server.SimpleHTTPRequestHandler
+        httpd = socketserver.TCPServer(("", port), handler)
+        thread = threading.Thread(target=httpd.serve_forever)
+        thread.daemon = True
+        thread.start()
+        return httpd
+
+    def test_render_config_contains_backup(self):
+        cfg = NginxConfig(name='test', server_name='example.com', primary_upstream='remote:8000', backup_upstream='127.0.0.1:8000')
+        text = cfg.render_config()
+        self.assertIn('backup', text)
+        self.assertIn('proxy_set_header Upgrade $http_upgrade;', text)
+
+    def test_connection(self):
+        server = self._run_server(8123)
+        try:
+            cfg = NginxConfig(name='test', server_name='example.com', primary_upstream='127.0.0.1:8123')
+            self.assertTrue(cfg.test_connection())
+            cfg.primary_upstream = '127.0.0.1:8999'
+            self.assertFalse(cfg.test_connection())
+        finally:
+            server.shutdown()
+            server.server_close()
